refactor(qualité): drainer sept entrées de plus, dont deux gardes mortes (#1482)
Baseline : 663 → 656 erreurs, 320 → 313 blocs, régénération comparée bloc à
bloc — aucune entrée ajoutée. `app/` descend à 46 erreurs masquées, contre
153 au début de la série.

**Deux gardes qui ne gardaient rien.**

`UserActivityChart` écrivait `($date = Carbon::create(...)) ? $date->format('M') : ''`.
Carbon 3 rend une instance ou lève — il ne rend jamais `false` comme Carbon 2
le faisait. La branche fausse était inatteignable. `create()` restant déclaré
nullable, le retirer seul n'aurait fait que déplacer le problème : `parse()`
ne l'est pas, et l'année est de toute façon arbitraire puisque seul le mois
est lu.

`DailyJournalResource` repliait `date` sur `null`. La colonne est NOT NULL et
porte la moitié d'un index unique. Le docblock de la ressource, qui
redéclare ses propriétés au lieu de suivre le modèle, avait dérivé et
annonçait `Carbon|null` — le modèle, lui, dit juste.

**Une variable de boucle qui en cachait une autre.**
`CreateWorkoutTemplateFromWorkoutAction` parcourt deux fois la même
collection, pour deux raisons distinctes, et réutilisait `$line` alors que
celui de la première boucle était encore en portée. La seconde s'appelle
désormais `$sourceLine`.

**Deux résultats de validation dont la forme est déclarée.**
`validate()` rend un `array` non typé, donc lire une clé dessus était un
accès sur du `mixed`.

Cette dernière a demandé deux passes, et c'est la partie qui mérite d'être
lue. Déclarer `array{per_page?: int}` rendait le cast vers `int` apparemment
inutile — sauf que la règle `integer` de Laravel **valide sans convertir** :
une valeur passée en paramètre d'URL arrive en chaîne. Retirer le cast a fait
répondre 500 à `SupplementLogSecurityTest`. La forme honnête est
`int|numeric-string`, et le cast reste. J'avais écrit exactement le genre de
docblock menteur que cette série passe son temps à retirer.

Plafond du ratchet : 313 blocs / 656 erreurs. 2029 tests passent.

+++ app/Http/Controllers/Api/DailyJournalController.php
@@ -57,12 +57,22 @@ class DailyJournalController extends Controller
     {
         $this->authorize('viewAny', DailyJournal::class);
 
+        /**
+         * La forme est declaree, pas supposee : `validate()` rend un `array`
+         * non type, donc lire une cle dessus etait un acces sur du `mixed`. La
+         * regle juste au-dessus garantit exactement cette forme.
+         *
+         * `int|numeric-string` et non `int` : la regle `integer` de Laravel
+         * VALIDE sans convertir, donc une valeur passee en parametre d'URL
+         * arrive en chaine.
+         *
+         * @var array{per_page?: int|numeric-string} $validated
+         */
         $validated = $request->validate([
             'per_page' => 'sometimes|integer|min:1|max:100',
         ]);
 
-        /** @var int $perPage */
-        $perPage = $validated['per_page'] ?? 15;
+        $perPage = (int) ($validated['per_page'] ?? 15);
 
         $journals = $this->user()->dailyJournals()
             ->orderBy('date', 'desc')
